Improved Header generator script

+++ class-dump-z_0.2a/mac_x86/generate_dump.sh
@@ -1,6 +1,43 @@
-clear
-FPATH=/Applications/Xcode45-DP2.app/Contents/Developer/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS6.0.sdk/System/Library/Frameworks/
-ROOTFOLDER=iPhoneOS6.0.sdk
+if [ -z $1 ]
+	then
+		echo ""
+		echo "Objective C Frameworks Header Dump Generator"
+		echo "Version: 1.0 Release Date: 2012-06-27"
+		echo ""
+		echo "Authors:"
+		echo "@loretoparisi @christian_zanin"
+		echo ""
+ 		echo "Usage:"
+		echo "generate_dump <framework root folder> <output folder>"
+		echo ""
+		exit
+fi
+
+if [ -z $2 ]
+	then
+		echo "Error. Please specifify the output folder"
+		echo "Usage:"
+                echo "generate_dump <framework root folder> <output folder>"
+                echo ""
+                exit
+fi
+
+if [ ! -d $1 ] 
+	then
+		echo ""
+		echo "Error. Framework directory does not exists"
+		echo "Generally frameworks are like the following"
+		echo "/Applications/Xcode.app/Contents/Developer/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS6.0.sdk/System/Library/Frameworks/"
+		echo ""
+		exit
+fi
+exit
+
+#FPATH=/Applications/Xcode45-DP2.app/Contents/Developer/Platforms/iPhoneOS.platform/Developer/SDKs/iPhoneOS6.0.sdk/System/Library/Frameworks/
+
+FPATH=$1
+ROOTFOLDER=$2
+
 if [ -d $ROOTFOLDER ]
 	then
  		rm $ROOTFOLDER
